src/requirements.txt
RPi.GPIO>=0.7,<2.0
evdev>=1.3.0,<2.0
